--- a/src/OptimizedEinsum.jl
+++ b/src/OptimizedEinsum.jl
@@ -1,11 +1,11 @@
 module OptimizedEinsum
+using Requires: @require
 
 include("Counters.jl")
 include("Utils.jl")
 include("ContractionPath.jl")
 include("Solvers/Solvers.jl")
 
-using Requires: @require
 function __init__()
     @require Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a" include("Visualization.jl")
 end