Fix #11. Now Makie is not loaded by default and instead is added as a `@require` from `Visualization.jl`.