@mofeing 2022-12-21
Nice! With this PR the loading time goes from 10.60s to 0.08s

### Before
```julia
julia> @time using OptimizedEinsum
 10.599981 seconds (38.09 M allocations: 2.434 GiB, 5.40% gc time, 2.96% compilation time: 65% of which was recompilation)
```

### After
```julia
julia> @time using OptimizedEinsum
  0.078971 seconds (168.63 k allocations: 12.916 MiB, 6.95% compilation time: 18% of which was recompilation)
```